refactor: 매장 관련 컴포넌트 폴더 정리 및 컴포넌트 분리 작업 #11
