+++ src/app/store-list/[store]/review/page.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import ReviewForm from '@/components/ReviewForm';
+import ReviewForm from '@/components/stores/ReviewForm';
 
 function Review() {
   return (
